requirements.txt
streamlit
streamlit_chat
sounddevice
cython
sentencepiece
soundfile
streamlit-audiorecorder
ffmpeg-python
google-cloud-speech==2.9.0
pydub==0.25.1
